Refactored UI code into separate file
